requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.3.0
django-createsuperuserwithpassword==2.0.0
django-localflavor==3.0.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-nested-routers==0.91
gunicorn==20.0.4
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.8.5
Pygments==2.6.1
PyJWT==1.7.1
python-dotenv==0.13.0
python-stdnum==1.13
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
